refactor: move services to api folder
