+++ src/api/userServise.ts
@@ -1,5 +1,5 @@
 import { AxiosResponse } from 'axios';
-import api from '../api';
+import api from '.';
 import { NewUser } from '../types/user';
 
 export default class UserService {
